fix: no force pino types dep (#6)
closes graphql-nexus/nexus#808
closes graphql-nexus/nexus#832

This also does a refactor wherein pino is no longer used. Although pino was used for performance reasons the numerous hacks to get it integrated actually mean the immediate result might be better performance (no benchmarks to back this up).

The refactor's primary goal was however to simplify the codebase.