feat: new syntax for inclusive/exclusive descedants
New changes make it easier to adopt the forthcoming singleton where
there will be an implicit root to all paths.

The problem was this:

We want `*` to match all. But the implicit root suggests that `*` is
shorthand for `.:*`. Example, before:

```
*
a*
a:*
```

As-is, would newly exapnd to:

```
.:*
.:a*
.:a:*
```

Being consistent would mean that `.:*` is exclusive descendants of root,
but not root. But very likely that is not what most users will want/mean
by `*`. What they would want most of the time is:

```
.*
```

But having to remember to write that would be bad design, making the
common thing harder than easier.

Therefore to better support the implicit root path syntax we made
exclusiion be expressed by layering on one more `:`. This means in the
future, it will be possible to write the awkward pattern of:

```
:*
```

Which will expand to `.::*` which will mean "all descedants of root but
not root itself aka. exclusive". This is a somewhat odd pattern and will
probably not be desired in most cases. Thus the added/awkward syntax to
achieve it is acceptable/proportional.

BREAKING CHANGE:

1. Descendant matcher must now be after `:`

   ```
   Before:

   foo*

   Now:

   foo:*
   ```

2. Descendants `:` matcher is inclusive now

3. Descendants exclusively syntax uses `::` now

   ```
   Before:

   foo:*

   Now:

   foo::*
   ```