feat(data): lazy context prop
This makes working with JSON logs more reasonable in a terminal. This
may cause problems in systems with poor support for nullability. In the
future we may bring back an option to have a strict context property.
Its default would probably be off in dev and on in prod.

BREAKING CHANGE:

1. When context is empty then the `context` property in the respective log record will be omitted.

      Before:

      ```
      log.info('hello')
      // { "context": {}, ... }
      ```

      Now:

      ```
      log.info('hello')
      // { ... }
      ```